job-agent/requirements.txt
anthropic>=0.40.0
requests>=2.32.0
python-dotenv>=1.0.0
schedule>=1.2.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
playwright>=1.40.0
pytest>=8.0.0
pytest-mock>=3.10.0
